Send initialized call after recieving server settings
